Issue #8: Added authorization methods (including rest api), implemented core data (tracks) & system (tick, check, auth) api.

+++ tales_django/entities/Users/views/delete_user_account.py
@@ -19,7 +19,8 @@ logger = getDebugLogger()
 def delete_user_account(request: HttpRequest):
     try:
         user = request.user
-        user.session_set.all().delete()
+        if hasattr(user, 'session_set'):
+            user.session_set.all().delete()
         user.delete()
         logout(request)
         messages.success(request, 'Your account has been removed successfully')
